+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 For more information about changelogs, check [Keep a Changelog](http://keepachangelog.com) and
 [Vandamme](http://tech-angels.github.io/vandamme).
 
+## Unreleased
+
+* [Fix] The footer reads `Displaying items 1-20 of 101`, without `in total`
+
 ## 4.4.0 - 2026-09-08
 
 * [Fix] Tapping the light/dark toggle or the log out on a phone leaves no tint behind